requirements.txt
PyQt6>=6.6.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.110.0
schedule>=1.2.0
python-dateutil>=2.8.0
